correction bug

+++ asm/src/parser/check_nbr_arg.c
@@ -6,6 +6,7 @@
 */
 
 #include <stdbool.h>
+#include <stdlib.h>
 #include "op.h"
 
 bool check_nbr_arg(char id, char *line)
@@ -20,7 +21,7 @@ bool check_nbr_arg(char id, char *line)
 			nb_arg++;
 		i++;
 	}
-	if (op_tab[id].nbr_args != nb_arg + 1)
+	if (op_tab[ (int) id].nbr_args != nb_arg + 1)
 		return (false);
 	return (true);
 }
